requirements.txt
aiogram==3.13.1
gspread==6.1.2
python-dotenv==1.0.1
oauth2client
aiohttp
